Unittest: make all integral types size_t
instead of a mix of size_t, int, and uint

+++ tests/unittest.h
@@ -1136,9 +1136,9 @@ static Vc::enable_if<!Vc::MIC::is_vector<Vec>::value, typename Vec::Mask> allMas
     const Vec indexes(Vc::IndexesFromZero);
     M mask(true);
 
-    for (int j = 0; j < int(8 * sizeof(i)); ++j) {
-        if (i & (1u << j)) {
-            mask ^= static_cast<M>(indexes == j);
+    for (size_t j = 0; j < 8 * sizeof(i); ++j) {
+        if (i & (size_t(1) << j)) {
+            mask ^= static_cast<M>(indexes == int(j));
         }
     }
     return mask;
